+++ plugin-gradle/src/main/java/dev/equo/ide/gradle/EquoIdeExtension.java
@@ -38,6 +38,10 @@ public class EquoIdeExtension extends P2ModelDsl {
 		ideHooks.add(branding);
 	}
 
+	public IdeHookBranding branding() {
+		return branding;
+	}
+
 	private IdeHookWelcome welcome = null;
 
 	public IdeHookWelcome welcome() {